[INTERNAL] Normalize path used for resource objects
Adjusts the existing Saros resource implementations to always normalize
the given paths. This ensures that Saros never operates on paths
pointing outside of the current workspace (without resorting to symlink
trickery). More specifically, it ensures that the relative paths
contained in the resource object can't be used to point to a parent
directory of the reference point once resolved.