[UI][I] Add option to add user to existing session
Adds the option to add a user to an existing session. This enables
multi-user session support for Saros/I.

Resolves issues where the pop-up dialog to create a new session was
still displayed, even if the local user is already part of a session.
Ensures that the dialog is only displayed when there is a valid option
and that the correct option is displayed.

Copies the icon used for the option to add users to the session from
the Eclipse package.

Resolves part of #1057.